refactor: change func name and file names
